USER DONE CHARGEABLE ON TOUCH

+++ app/src/main/java/com/shivam/app_ka_kaam/Java_objects/electricityconstants.java
@@ -1,41 +1,41 @@
 package com.shivam.app_ka_kaam.Java_objects;
 
 public class electricityconstants {
-    int c1;
-    int c2;
-    int c3;
+    double c1;
+    double c2;
+    double c3;
 
     public electricityconstants(){
 
     }
 
-    public electricityconstants(int c1, int c2, int c3) {
+    public electricityconstants(double c1, double c2, double c3) {
         this.c1 = c1;
         this.c2 = c2;
         this.c3 = c3;
     }
 
-    public int getC1() {
+    public double getC1() {
         return c1;
     }
 
-    public void setC1(int c1) {
+    public void setC1(double c1) {
         this.c1 = c1;
     }
 
-    public int getC2() {
+    public double getC2() {
         return c2;
     }
 
-    public void setC2(int c2) {
+    public void setC2(double c2) {
         this.c2 = c2;
     }
 
-    public int getC3() {
+    public double getC3() {
         return c3;
     }
 
-    public void setC3(int c3) {
+    public void setC3(double c3) {
         this.c3 = c3;
     }
 }
